+++ CHANGELOG.md
@@ -5,3 +5,5 @@ This log is sorted from oldest to newest, oldest first.
 ### 14.2 Getting Started
 Builds clean using `tcc -o bin/clox src/*.c`.  I'm sure this'll fail later.
 
+### 14.3 Chunks of Instructions
+Bytecode chunks and dynamic array resizing.  Compiles and "runs."